+++ fast_nav/test/fast_nav_test.dart
@@ -3,16 +3,11 @@ import 'dart:async';
 import 'package:fast_nav/fast_nav.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'observer_tester.dart';
 
 Widget buildBaseWidget() => MaterialApp(
       navigatorKey: FastNav.init(),
-      navigatorObservers: [FastNavObserver()],
-      onGenerateRoute: (settings) => FastNav.generateAnonymousRoute(
-        settings: settings,
-        page: const Text('home'),
-      ),
       routes: {
+        '/': (context) => const Text('home'),
         'test_page': (context) => const Text('test_page'),
         'test_page_2': (context) => const Text('test_page_2'),
         'test_page_3': (context) => const Text('test_page_3'),
@@ -21,12 +16,9 @@ Widget buildBaseWidget() => MaterialApp(
     );
 
 Widget buildNestedNavigatorWidget() => MaterialApp(
-      onGenerateRoute: (settings) => FastNav.generateAnonymousRoute(
-        settings: settings,
-        page: NestedNavigator(
-          name: 'nestedNavigator',
-          home: const Text('home'),
-        ),
+      home: NestedNavigator(
+        name: 'nestedNavigator',
+        home: const Text('home'),
       ),
     );
 
@@ -45,50 +37,6 @@ void main() {
       expect((e as NavigatorNotRegistered).navigatorName, 'test');
     }
 
-    await tester.pumpWidget(
-      MaterialApp(
-        navigatorKey: FastNav.init(),
-        home: const Text('home'),
-      ),
-    );
-    try {
-      unawaited(FastNav.push(const SizedBox.shrink(), preventDuplicates: true));
-      throw 'Did not throw';
-    } catch (e) {
-      expect((e as NavigatorObserverNotRegistered).navigatorName, isNull);
-    }
-
-    // Not using FastNav.reset() simulates a hot reload
-    await tester.pumpWidget(
-      MaterialApp(
-        navigatorKey: FastNav.init(),
-        home: const Text('home'),
-        navigatorObservers: [FastNavObserver()],
-      ),
-    );
-    try {
-      unawaited(FastNav.push(const SizedBox.shrink(), preventDuplicates: true));
-      throw 'Did not throw';
-    } catch (e) {
-      expect((e as NavigatorObserverNotRegistered).navigatorName, isNull);
-    }
-
-    FastNav.reset();
-
-    await tester.pumpWidget(
-      MaterialApp(
-        navigatorKey: FastNav.init(),
-        home: const Text('home'),
-        navigatorObservers: [FastNavObserver()],
-      ),
-    );
-    try {
-      unawaited(FastNav.push(const SizedBox.shrink(), preventDuplicates: true));
-      throw 'Did not throw';
-    } catch (e) {
-      expect((e as AnonymousRouteNotPatched).navigatorName, isNull);
-    }
-
     FastNav.reset();
 
     await tester.pumpWidget(
@@ -104,7 +52,6 @@ void main() {
       FastNav.push(
         const SizedBox.shrink(),
         navigatorName: 'nestedNavigator',
-        preventDuplicates: true,
       ),
     );
 
@@ -163,34 +110,6 @@ void main() {
     expect(FastNav.canPop(), false);
   });
 
-  testWidgets('Anonymous duplicate prevention', (tester) async {
-    await tester.pumpWidget(buildBaseWidget());
-
-    unawaited(FastNav.push(const Text(''), preventDuplicates: true));
-    await tester.pumpAndSettle();
-    expect(find.text('home'), findsOneWidget);
-
-    unawaited(FastNav.pushReplacement(const Text(''), preventDuplicates: true));
-    await tester.pumpAndSettle();
-    expect(find.text('home'), findsOneWidget);
-
-    unawaited(
-      FastNav.pushAndRemoveUntil(
-        const Text(''),
-        (e) => false,
-        preventDuplicates: true,
-      ),
-    );
-    await tester.pumpAndSettle();
-    expect(find.text('home'), findsOneWidget);
-
-    unawaited(
-      FastNav.pushAndRemoveAll(const Text(''), preventDuplicates: true),
-    );
-    await tester.pumpAndSettle();
-    expect(find.text('home'), findsOneWidget);
-  });
-
   testWidgets('Named routes', (tester) async {
     await tester.pumpWidget(buildBaseWidget());
     expect(FastNav.canPop(), false);
@@ -237,53 +156,6 @@ void main() {
     expect(FastNav.canPop(), false);
   });
 
-  // Based on official Flutter navigation tests
-  testWidgets('Named duplicate prevention', (tester) async {
-    final testObserver = TestObserver();
-
-    await tester.pumpWidget(
-      MaterialApp(
-        home: const Text('home'),
-        navigatorKey: FastNav.init(),
-        routes: {'test_page': (context) => const Text('test_page')},
-        navigatorObservers: [FastNavObserver(), testObserver],
-      ),
-    );
-    expect(find.text('home'), findsOneWidget);
-
-    // Push a named route
-    unawaited(FastNav.pushNamed('test_page'));
-    await tester.pumpAndSettle();
-    expect(find.text('test_page'), findsOneWidget);
-
-    var isPushed = false;
-    testObserver.onPushed = (route, previousRoute) => isPushed = true;
-
-    unawaited(FastNav.pushNamed('test_page', preventDuplicates: true));
-    await tester.pumpAndSettle();
-
-    unawaited(
-      FastNav.pushReplacementNamed('test_page', preventDuplicates: true),
-    );
-    await tester.pumpAndSettle();
-
-    unawaited(
-      FastNav.pushNamedAndRemoveUntil(
-        'test_page',
-        (e) => false,
-        preventDuplicates: true,
-      ),
-    );
-    await tester.pumpAndSettle();
-
-    unawaited(
-      FastNav.pushNamedAndRemoveAll('test_page', preventDuplicates: true),
-    );
-    await tester.pumpAndSettle();
-
-    expect(isPushed, isFalse);
-  });
-
   testWidgets('NestedNavigator', (tester) async {
     await tester.pumpWidget(buildNestedNavigatorWidget());
     expect(find.text('home'), findsOneWidget);
@@ -302,51 +174,5 @@ void main() {
     await tester.pumpAndSettle();
     expect(find.text('home'), findsOneWidget);
     expect(FastNav.canPop(navigatorName: 'nestedNavigator'), false);
-  });
-
-  testWidgets('NestedNavigator duplicate prevention', (tester) async {
-    await tester.pumpWidget(buildNestedNavigatorWidget());
-    expect(find.text('home'), findsOneWidget);
-
-    unawaited(
-      FastNav.push(
-        const Text(''),
-        navigatorName: 'nestedNavigator',
-        preventDuplicates: true,
-      ),
-    );
-    await tester.pumpAndSettle();
-    expect(find.text('home'), findsOneWidget);
-
-    unawaited(
-      FastNav.pushReplacement(
-        const Text(''),
-        navigatorName: 'nestedNavigator',
-        preventDuplicates: true,
-      ),
-    );
-    await tester.pumpAndSettle();
-    expect(find.text('home'), findsOneWidget);
-
-    unawaited(
-      FastNav.pushAndRemoveUntil(
-        const Text(''),
-        (e) => false,
-        navigatorName: 'nestedNavigator',
-        preventDuplicates: true,
-      ),
-    );
-    await tester.pumpAndSettle();
-    expect(find.text('home'), findsOneWidget);
-
-    unawaited(
-      FastNav.pushAndRemoveAll(
-        const Text(''),
-        navigatorName: 'nestedNavigator',
-        preventDuplicates: true,
-      ),
-    );
-    await tester.pumpAndSettle();
-    expect(find.text('home'), findsOneWidget);
   });
 }